Refactored header component
