+++ src/components/header/export-modal/index.tsx
@@ -1,6 +1,4 @@
 import { connect } from 'react-redux';
-import { bindActionCreators, Dispatch } from 'redux';
-import * as EditorActions from '../../../actions/editor';
 import { State } from '../../../constants/default-state';
 import Renderer from './renderer';
